backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
anthropic==0.42.0
asyncpg==0.30.0
psycopg2-binary==2.9.10
pydantic==2.10.4
pydantic-settings==2.7.1
python-dotenv==1.0.1
beautifulsoup4==4.12.3
requests==2.32.3
sentence-transformers==3.3.1
numpy==2.2.2
sse-starlette==2.2.1
